chore(v3): refactor provision and deploy utils (#5915)
* feat: up

* fix: ut

* feat: up

* fix: ut

* fix: error

* fix: error

* fix: ut

* feat: up

* fix: up
