+++ packages/fx-core/tests/component/workflow.test.ts
@@ -25,8 +25,6 @@ import "../../src/component/feature/sql";
 import "../../src/component/resource/botService";
 import { createContextV3 } from "../../src/component/utils";
 import { deleteFolder, MockTools, randomAppName } from "../core/utils";
-import * as provisionV3 from "../../src/plugins/solution/fx-solution/v3/provision";
-import * as provisionV2 from "../../src/plugins/solution/fx-solution/v2/provision";
 import { AppStudioClient } from "../../src/plugins/resource/appstudio/appStudio";
 import * as clientFactory from "../../src/plugins/resource/bot/clientFactory";
 import { AADRegistration } from "../../src/plugins/resource/bot/aadRegistration";
@@ -63,9 +61,10 @@ import { apiConnectorImpl } from "../../src/component/feature/apiConnector";
 import * as backup from "../../src/plugins/solution/fx-solution/utils/backupFiles";
 import { AadApp } from "../../src/component/resource/aadApp/aadApp";
 import { Constants } from "../../src/plugins/resource/aad/constants";
-import * as deployV3 from "../../src/plugins/solution/fx-solution/v3/deploy";
 import { CoreQuestionNames } from "../../src/core/question";
 import * as questionV3 from "../../src/component/questionV3";
+import { provisionUtils } from "../../src/component/provisionUtils";
+import { deployUtils } from "../../src/component/deployUtils";
 describe("Workflow test for v3", () => {
   const sandbox = sinon.createSandbox();
   const tools = new MockTools();
@@ -292,9 +291,9 @@ describe("Workflow test for v3", () => {
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
     sandbox
-      .stub(provisionV3, "fillInAzureConfigs")
+      .stub(provisionUtils, "fillInAzureConfigs")
       .resolves(ok({ hasSwitchedSubscription: false }));
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
     sandbox.stub(clientFactory, "createResourceProviderClient").resolves({});
@@ -373,8 +372,10 @@ describe("Workflow test for v3", () => {
     sandbox
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
-    sandbox.stub(provisionV3, "fillInAzureConfigs").resolves(ok({ hasSwitchedSubscription: true }));
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox
+      .stub(provisionUtils, "fillInAzureConfigs")
+      .resolves(ok({ hasSwitchedSubscription: true }));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
     sandbox.stub(clientFactory, "createResourceProviderClient").resolves({});
@@ -596,8 +597,10 @@ describe("Workflow test for v3", () => {
     sandbox
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
-    sandbox.stub(provisionV3, "fillInAzureConfigs").resolves(ok({ hasSwitchedSubscription: true }));
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox
+      .stub(provisionUtils, "fillInAzureConfigs")
+      .resolves(ok({ hasSwitchedSubscription: true }));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
     sandbox.stub(clientFactory, "createResourceProviderClient").resolves({});
@@ -678,10 +681,10 @@ describe("Workflow test for v3", () => {
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
     sandbox
-      .stub(provisionV3, "fillInAzureConfigs")
+      .stub(provisionUtils, "fillInAzureConfigs")
       .resolves(ok({ hasSwitchedSubscription: false }));
     sandbox
-      .stub(provisionV2, "askForProvisionConsentNew")
+      .stub(provisionUtils, "askForProvisionConsentNew")
       .resolves(err(new UserError("Solution", "CancelProvision", "CancelProvision")));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
@@ -789,8 +792,10 @@ describe("Workflow test for v3", () => {
     sandbox
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
-    sandbox.stub(provisionV3, "fillInAzureConfigs").resolves(ok({ hasSwitchedSubscription: true }));
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox
+      .stub(provisionUtils, "fillInAzureConfigs")
+      .resolves(ok({ hasSwitchedSubscription: true }));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(backup, "backupFiles").resolves(ok(undefined));
     sandbox
       .stub(armFunctions, "updateAzureParameters")
@@ -859,7 +864,7 @@ describe("Workflow test for v3", () => {
     sandbox
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
     sandbox.stub(clientFactory, "createResourceProviderClient").resolves({});
@@ -920,7 +925,7 @@ describe("Workflow test for v3", () => {
     sandbox
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
     sandbox.stub(clientFactory, "createResourceProviderClient").resolves({});
@@ -1041,9 +1046,9 @@ describe("Workflow test for v3", () => {
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
     sandbox
-      .stub(provisionV3, "fillInAzureConfigs")
+      .stub(provisionUtils, "fillInAzureConfigs")
       .resolves(ok({ hasSwitchedSubscription: false }));
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
     sandbox.stub(clientFactory, "createResourceProviderClient").resolves({});
@@ -1054,7 +1059,7 @@ describe("Workflow test for v3", () => {
       objectId: "00000000-0000-0000-0000-000000000000",
     });
     sandbox.stub(arm, "deployArmTemplates").resolves(ok(undefined));
-    sandbox.stub(deployV3, "checkDeployAzureSubscription").resolves(ok({}));
+    sandbox.stub(deployUtils, "checkDeployAzureSubscription").resolves(ok({}));
     sandbox.stub(AadApp.prototype, "provision").resolves(ok(undefined));
     sandbox.stub(AadApp.prototype, "setApplicationInContext").resolves(ok(undefined));
     sandbox.stub(AadApp.prototype, "configure").resolves(ok(undefined));
@@ -1150,9 +1155,9 @@ describe("Workflow test for v3", () => {
       .stub(tools.tokenProvider.azureAccountProvider, "getAccountCredentialAsync")
       .resolves(TestHelper.fakeCredential);
     sandbox
-      .stub(provisionV3, "fillInAzureConfigs")
+      .stub(provisionUtils, "fillInAzureConfigs")
       .resolves(ok({ hasSwitchedSubscription: false }));
-    sandbox.stub(provisionV2, "askForProvisionConsentNew").resolves(ok(Void));
+    sandbox.stub(provisionUtils, "askForProvisionConsentNew").resolves(ok(Void));
     sandbox.stub(AppStudioClient, "getApp").onFirstCall().throws({}).onSecondCall().resolves({});
     sandbox.stub(AppStudioClient, "importApp").resolves({ teamsAppId: "mockTeamsAppId" });
     sandbox.stub(clientFactory, "createResourceProviderClient").resolves({});
@@ -1163,7 +1168,7 @@ describe("Workflow test for v3", () => {
       objectId: "00000000-0000-0000-0000-000000000000",
     });
     sandbox.stub(arm, "deployArmTemplates").resolves(ok(undefined));
-    sandbox.stub(deployV3, "checkDeployAzureSubscription").resolves(ok({}));
+    sandbox.stub(deployUtils, "checkDeployAzureSubscription").resolves(ok({}));
     sandbox.stub(AadApp.prototype, "provision").resolves(ok(undefined));
     sandbox.stub(AadApp.prototype, "setApplicationInContext").resolves(ok(undefined));
     sandbox.stub(AadApp.prototype, "configure").resolves(ok(undefined));
