+++ packages/fx-core/tests/plugins/solution/solution.provision.test.ts
@@ -102,8 +102,7 @@ import * as backup from "../../../src/plugins/solution/fx-solution/utils/backupF
 import { resourceGroupHelper } from "../../../src/plugins/solution/fx-solution/utils/ResourceGroupHelper";
 import * as manifestTemplate from "../../../src/plugins/resource/appstudio/manifestTemplate";
 import { SolutionRunningState } from "../../../src/plugins/solution/fx-solution/types";
-import mockedEnv from "mocked-env";
-import { askForProvisionConsentNew } from "../../../src/plugins/solution/fx-solution/v2/provision";
+import { provisionUtils } from "../../../src/component/provisionUtils";
 
 chai.use(chaiAsPromised);
 const expect = chai.expect;
@@ -1470,7 +1469,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    await askForProvisionConsentNew(
+    await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1538,7 +1537,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1608,7 +1607,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1678,7 +1677,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1748,7 +1747,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1800,7 +1799,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1846,7 +1845,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1910,7 +1909,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
@@ -1971,7 +1970,7 @@ describe("askForProvisionConsentNew", () => {
     const m365AccountName = "m365Name";
 
     // Act
-    const res = await askForProvisionConsentNew(
+    const res = await provisionUtils.askForProvisionConsentNew(
       mockedCtx,
       mockedAzureTokenProvider,
       mockedEnvInfo as v3.EnvInfoV3,
